requirements.txt
selenium==4.18.1
pytest==8.0.2
pytest-html==4.1.1
pytest-xdist==3.5.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
allure-pytest==2.13.2
faker==24.1.0
n